# docs/swagger.yaml
basePath: /api/v1
definitions:
  chat_sdk.CommentMomentReq:
    properties:
      content:
        type: string
      moment_id:
        type: integer
      parent_id:
        type: integer
    required:
    - content
    - moment_id
    type: object
...
  chat_sdk.ForwardMessageReq:
    properties:
      comment:
        type: string
      items:
        items:
          properties:
            message_id:
              type: integer
          type: object
        type: array
      mode:
        description: merge/single
        example: merge
        type: string
      to_room_ids:
        items:
          type: integer
        type: array
    required:
    - items
    - to_room_ids
    type: object
...
  response.Response:
    properties:
      code:
        description: 业务状态码
        example: 0
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 提示消息
        example: success
        type: string
    type: object
  service.CommentDTO:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      moment_id:
        type: integer
      parent_id:
        type: integer
      user_id:
        type: integer
    type: object
  service.ConversationListItemDTO:
    properties:
      avatar:
        description: 私聊：对方头像；群聊：群头像
        type: string
      conversation_id:
        type: integer
      last_message:
        $ref: '#/definitions/service.MessageDTO'
      name:
        description: 私聊：对方昵称；群聊：群名
        type: string
      room_account:
        type: string
      room_id:
        type: integer
      room_type:
        description: 1-私聊 2-群聊
        type: integer
      unread_count:
        type: integer
      updated_at:
        description: unix seconds for easy sort/render
        type: integer
      user_id:
        description: 私聊时为对方用户ID，群聊时为0
        type: integer
    type: object
  service.CreateMomentReq:
    properties:
      images:
        description: 最多9张
        items:
          type: string
        type: array
      title:
        type: string
      video:
        description: 单个视频URL
        type: string
    type: object
...
  service.MessageDTO:
    properties:
      content:
        type: string
      created_at:
        type: string
      extra:
        items:
          type: integer
        type: array
      id:
        type: integer
      is_encrypted:
        type: boolean
      is_system:
        type: boolean
      message_id:
        type: string
      reply_to_msg_id:
        type: integer
      room_id:
        type: integer
      sender_id:
        type: integer
      status:
        type: integer
      type:
        type: integer
      updated_at:
        type: string
    type: object
  service.MessageListItemDTO:
    properties:
      content:
        type: string
      created_at:
        type: string
      extra:
        items:
          type: integer
        type: array
      id:
        type: integer
      is_encrypted:
        type: boolean
      is_system:
        type: boolean
      reply_to_msg_id:
        type: integer
      room_id:
        type: integer
      sender:
        $ref: '#/definitions/service.SenderDTO'
      sender_id:
        type: integer
      status:
        type: integer
      type:
        type: integer
      updated_at:
        type: string
    type: object
  service.MomentDTO:
    properties:
      comments_cnt:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      images_count:
        type: integer
      media_type:
        type: integer
      medias:
        items:
          $ref: '#/definitions/service.MomentMediaDTO'
        type: array
      title:
        type: string
      user_id:
        type: integer
    type: object
  service.MomentMediaDTO:
    properties:
      sort:
        type: integer
      type:
        description: 1-图片 2-视频
        type: integer
      url:
        type: string
    type: object
...
  service.SenderDTO:
    properties:
      avatar:
        type: string
      id:
        type: integer
      nickname:
        type: string
      username:
        type: string
    type: object
...
host: localhost:6789
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cydxin/chat-sdk/issues
  description: |-
    即时通讯 SDK 的 RESTful API 文档，包含用户管理、消息、房间、好友等模块

    ## 业务状态码说明
    | Code | 说明 |
    |------|------|
    | 0 | 成功 |
    | 10001 | 参数错误 |
    | 10002 | 用户不存在 |
    | 10003 | 密码错误（登录失败） |
    | 10004 | Token 无效 |
    | 10005 | 权限不足 |
    | 99999 | 内部错误 |

    ## HTTP 状态码说明
    - **200**: 业务请求成功（根据 response.code 判断业务状态）
    - **401**: 认证失败（未登录/Token 无效/登录失败）
    - **403**: 权限不足
    - **500**: 服务器内部错误

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
    "code": 0,
    "msg": "success",
    "data": {}
    }
    ```
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/cydxin/chat-sdk
  title: Chat SDK API
  version: "1.0"
paths:
...
  /message/conversation/hide:
    post:
      consumes:
      - application/json
      description: 将当前用户某个房间的会话从消息列表隐藏（仅影响自己；新消息会自动重新展示）
      parameters:
      - description: 房间ID
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 隐藏会话
      tags:
      - 消息
  /message/conversations:
    get:
      consumes:
      - application/json
      description: 获取当前用户的会话列表（未删除的会话），包含头像、名称、room、最后一条消息、未读数
      produces:
      - application/json
      responses:
        "200":
          description: 会话列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.ConversationListItemDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取消息列表
      tags:
      - 消息
  /message/detail:
    get:
      consumes:
      - application/json
      description: 根据消息ID获取消息详情
      parameters:
      - description: 消息ID
        format: int64
        in: query
        name: message_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 消息详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MessageDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取消息详情
      tags:
      - 消息
  /message/forward:
    post:
      consumes:
      - application/json
      description: 支持逐条转发(single) 或 合并转发(merge)
      parameters:
      - description: 转发请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.ForwardMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建的消息ID列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 转发消息
      tags:
      - 消息
  /message/list:
    get:
      consumes:
      - application/json
      description: 分页获取房间历史消息
      parameters:
      - description: 房间ID
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量 以你要查询的ID为基准，向前查询，不传则向后
        in: query
        name: mess_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 消息列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.MessageListItemDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取房间消息
      tags:
      - 消息
  /message/recall:
    post:
      consumes:
      - application/json
      description: 撤回指定消息
      parameters:
      - description: 消息ID
        format: int64
        in: query
        name: message_id
        required: true
        type: integer
      - description: '4-撤回（会在聊天窗口留下痕迹） 5-删除（自己不可见） 6/7-双删（Sender/非Sender删除)在私聊中互相可以删除，但在群中你只能删除自己的，已经管理员进行删除 '
        format: int32
        in: query
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 撤回消息
      tags:
      - 消息
  /moment/comment:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论内容
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.CommentMomentReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 评论动态
      tags:
      - 朋友圈
  /moment/comment/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 动态ID
        format: int64
        in: query
        name: moment_id
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 评论列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.CommentDTO'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取动态评论
      tags:
      - 朋友圈
  /moment/create:
    post:
      consumes:
      - application/json
      description: 标题 + 图片(最多9张) 或 视频(1个)
      parameters:
      - description: 动态内容（title, images(最多9) 或 video 二选一）
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.CreateMomentReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MomentDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 发布朋友圈动态
      tags:
      - 朋友圈
  /moment/list:
    get:
      consumes:
      - application/json
      description: 获取自己与好友发布的动态（按时间倒序）
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 动态列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.MomentDTO'
                  type: array
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 朋友圈动态列表
      tags:
      - 朋友圈
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  QueryToken:
    description: 用于 WebSocket 等无法传 header 的场景
    in: query
    name: token
    type: apiKey
swagger: "2.0"